Update the code with the new: - Complete unit tét for the repository package

+++ src/main/java/vn/com/fecredit/app/dto/participant/UpdateParticipantRequest.java
@@ -26,7 +26,7 @@ public class UpdateParticipantRequest {
 
     @NotNull
     @Min(1)
-    private Integer dailySpinLimit;
+    private Long dailySpinLimit;
 
     @NotNull
     private Boolean isActive;
